requirements.txt
streamlit
yfinance
pandas
plotly
python-dotenv
fredapi
mftool
cot_reports